Рефакторинг кода

+++ CoinMarketCup/Service/СoinMarketCupService.cs
@@ -10,6 +10,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CoinMarketCup.Helpers;
 
 
 namespace CoinMarketCup.Service
@@ -34,14 +35,17 @@ namespace CoinMarketCup.Service
         {
             if (await _coinMarketRepository.IsExpiryDateExpired())
             {
+                var cryptocurrencies = await GetCryptocurrencies();
+                
+                if (!cryptocurrencies.IsSuccessfully)
+                {
+                    _logger.LogInformation("Error get cryptocurrencies");
+                    return Return<List<Cryptocurrency>>.ReturnFail(cryptocurrencies.Error);
+                }
+
                 await _coinMarketRepository.DeleteAllDate();
 
-                bool isRecord = await DateRecord();
-
-                if (!isRecord)
-                {
-                    return Return<List<Cryptocurrency>>.ReturnFail("record_fail");
-                }
+                await DateRecord(cryptocurrencies.Information);
             }
 
             var result = await _coinMarketRepository.GetCryptocurrencies(paginatorInfo, sortState);
@@ -50,18 +54,9 @@ namespace CoinMarketCup.Service
 
         }
 
-
-        public async Task<bool> DateRecord()
+        public async Task DateRecord(List<Cryptocurrency> cryptocurrencies)
         {
-            var cryptocurrencies = await GetCryptocurrencies();
-
-            if (!cryptocurrencies.IsSuccessfully)
-            {
-                return false;
-            }
-            await _coinMarketRepository.AddRange(cryptocurrencies.Information);
-
-            return true;
+            await _coinMarketRepository.AddRange(cryptocurrencies);
         }
 
         public async Task<Return<List<Cryptocurrency>>> GetCryptocurrencies()
@@ -76,9 +71,12 @@ namespace CoinMarketCup.Service
             var listId = await GetIdCoin(listingLatestRequest.Information);
 
             var metadataRequest = await GetMetadata(listId);
+           
+            string fiatValue = await _settingCryptocurrencyRepository.GetFiatCurrency();
 
-            return !metadataRequest.IsSuccessfully ? Return<List<Cryptocurrency>>.ReturnFail(listingLatestRequest.Error)
-                : Return<List<Cryptocurrency>>.ReturnSuccessfully(await ObjectShapingCryptocurrencies(listingLatestRequest.Information, metadataRequest.Information));
+            return !metadataRequest.IsSuccessfully 
+                ? Return<List<Cryptocurrency>>.ReturnFail(listingLatestRequest.Error)
+                : Return<List<Cryptocurrency>>.ReturnSuccessfully(CoinMarketCupHelpers.ObjectShapingCryptocurrencies(listingLatestRequest.Information, metadataRequest.Information, fiatValue));
         }
 
         private async Task<Return<MetadataRequest>> GetMetadata(IEnumerable<string> listId)
@@ -98,7 +96,7 @@ namespace CoinMarketCup.Service
                 return Return<MetadataRequest>.ReturnFail("error_get_metadata");
             }
 
-            return Return<MetadataRequest>.ReturnSuccessfully(MetadataListToMetadata(metadataRequests));
+            return Return<MetadataRequest>.ReturnSuccessfully(CoinMarketCupHelpers.MetadataListToMetadata(metadataRequests));
         }
 
         private async Task<Return<ListingLatestRequest>> GetListingLatest()
@@ -115,66 +113,17 @@ namespace CoinMarketCup.Service
             }
 
         }
-        
+
         private async Task<IEnumerable<string>> GetIdCoin(ListingLatestRequest listingLatestRequest)
         {
             int counter = await _settingCryptocurrencyRepository.GetCountMetadata();
-            int i = await _settingCryptocurrencyRepository.GetCountMetadata();
+            int countMetadata = await _settingCryptocurrencyRepository.GetCountMetadata();
 
-            StringBuilder stringBuilder = new StringBuilder();
-            var result = new List<string>();
-
-            var listId = listingLatestRequest.Data
+            return  listingLatestRequest.Data
                 .Select(x => x.Id.ToString())
-                .GroupBy(w => counter++ / i)
+                .GroupBy(w => counter++ / countMetadata)
                 .Select(w => w.ToArray())
-                .ToList();
-
-            foreach (var id in listId)
-            {
-                stringBuilder.Clear();
-                stringBuilder.AppendJoin(',', id);
-                result.Add(stringBuilder.ToString());
-            }
-
-            return result;
-        }
-
-        private async Task<List<Cryptocurrency>> ObjectShapingCryptocurrencies(ListingLatestRequest listingLatestRequest, MetadataRequest metadataRequest)
-        {
-            string fiatValue = await _settingCryptocurrencyRepository.GetFiatCurrency();
-
-            return listingLatestRequest.Data.Select(item => new Cryptocurrency()
-            {
-                CoinMarketCupId = item.Id,
-                Name = item.Name,
-                Symbol = item.Symbol,
-                Logo = metadataRequest.Data[item.Id.ToString()].Logo,
-                LastUpdated = item.Quote[fiatValue].LastUpdated,
-                MarketCap = item.Quote[fiatValue].MarketCap,
-                PercentChange1H = item.Quote[fiatValue].PercentChange1H,
-                PercentChange24H = item.Quote[fiatValue].PercentChange24H,
-                Price = item.Quote[fiatValue].Price
-            })
-                .ToList();
-        }
-
-        private static MetadataRequest MetadataListToMetadata(IReadOnlyCollection<MetadataRequest> metadataRequests)
-        {
-            var metadataRequestResult = new MetadataRequest()
-            {
-                Data = new Dictionary<string, CryptoCurrencyMetadataRequest>(),
-                Status = metadataRequests.LastOrDefault()?.Status
-            };
-
-            foreach (var (id, cryptoCurrencyMetadata) in metadataRequests.SelectMany(metadataRequest => metadataRequest.Data))
-            {
-                metadataRequestResult
-                    .Data
-                    .Add(id, cryptoCurrencyMetadata);
-            }
-
-            return metadataRequestResult;
+                .Select(w => new StringBuilder().AppendJoin(',', w).ToString());
         }
 
     }
